quantity type of cart fixed in order entity

+++ internal/core/domain/order_product.go
@@ -11,7 +11,7 @@ type OrderProduct struct {
 	ID        uuid.UUID
 	OrderID   uuid.UUID
 	ProductID uuid.UUID
-	Quantity  int64
+	Quantity  uint8
 	CreatedAt time.Time
 	UpdatedAt time.Time
 
@@ -20,7 +20,7 @@ type OrderProduct struct {
 	Items *Product
 }
 
-func NewOrderProduct(orderID, productID uuid.UUID, quantity int64) *OrderProduct {
+func NewOrderProduct(orderID, productID uuid.UUID, quantity uint8) *OrderProduct {
 	return &OrderProduct{
 		ID:        uuid.Nil, // repository will asign the id
 		OrderID:   orderID,
